+++ changelog.md
@@ -1,4 +1,4 @@
-## Create: Alloyed v1.3d <br/> 
+## Create: Alloyed v1.4a <br/> 
 ### Fixed: <br/> 
-- A lot of blocks being unbreakable <br/> 
+- Create: Deco compat being broken <br/> 
 #### If you find an issue with the mod, please submit it&nbsp;<a href="https://github.com/MythrilBagels/Create-Alloyed/issues" rel="nofollow">here</a>.